requirements.txt
asgiref
dj-database-url
Django
django-cors-headers
django-filter
djangorestframework
djangorestframework-simplejwt
drf-yasg
gunicorn
inflection
packaging
Pillow
psycopg2-binary
PyJWT
python-dotenv
pytz
PyYAML
sqlparse
typing_extensions
tzdata
uritemplate
whitenoise
